fix(store): resolve LBA overflow and refactor kv_bench imports

Created-by: zhu-wentao40
Commit-by: Enigmo
Merged-by: Enigmo-x
Description: ## Description

<!-- A clear and concise description of the changes. Link to any relevant issues. -->

## Type of Change

* Types
  - [ ] Bug fix
  - [ ] New feature
    - [ ] Transfer Engine
    - [ ] Mooncake Store
    - [ ] Mooncake EP
    - [ ] Integration
    - [ ] P2P Store
    - [ ] Python Wheel
  - [ ] Breaking change
  - [ ] CI/CD
  - [ ] Documentation update
  - [ ] Other

## How Has This Been Tested?

<!-- Please describe the tests you've run to verify your changes. -->

## Checklist

- [ ] I have performed a self-review of my own code.
- [ ] I have formatted my own code using `./scripts/code_format.sh` before submitting.
- [ ] I have updated the documentation.
- [ ] I have added tests to prove my changes are effective.


See merge request: boostkit/mooncake!21

+++ mooncake-store/src/transfer_task.cpp
@@ -461,10 +461,10 @@ void SpdkNofWorkerPool::workerThread(int work_idx) {
                     SpdkNofTask *task = nof_qos->head[i];
                     SpdkNofSubTask *sub_task;
                     while (task->remaining_lba > 0 && avail_blocks > 0) {
-                        int submit_lba_count = std::min(avail_blocks, 
+                        uint32_t submit_lba_count = std::min(avail_blocks, 
                             std::min(task->remaining_lba, nof_qos->blocks_per_chunk));
                         int lba_off = task->lba_count - task->remaining_lba;
-                        int submit_lba= task->lba + lba_off;
+                        uint64_t submit_lba= task->lba + lba_off;
                         void *submit_ptr = reinterpret_cast<void *>
                             (reinterpret_cast<char *>(task->ptr) + lba_off * block_size);
